float type to clamping for blend

+++ gl_main.c
@@ -1437,9 +1437,9 @@ void R_SetupAliasFrame (entity_t *e, aliashdr_t *paliashdr, lerpdata_t *lerpdata
 	// always lerp
 	{
 		if (e->lerpflags & LERP_FINISH && numposes == 1)
-			lerpdata->blend = CLAMP (0, (cl.time - e->lerpstart) / (e->lerpfinish - e->lerpstart), 1);
+			lerpdata->blend = CLAMP (0.f, (float)(cl.time - e->lerpstart) / (e->lerpfinish - e->lerpstart), 1.f);
 		else
-			lerpdata->blend = CLAMP (0, (cl.time - e->lerpstart) / e->lerptime, 1);
+			lerpdata->blend = CLAMP (0.f, (float)(cl.time - e->lerpstart) / e->lerptime, 1.f);
 		lerpdata->pose1 = e->previouspose;
 		lerpdata->pose2 = e->currentpose;
 	}
@@ -1489,9 +1489,9 @@ void R_SetupEntityTransform (entity_t *e, lerpdata_t *lerpdata)
 	if (e != &cl.viewent && e->lerpflags & LERP_MOVESTEP)
 	{
 		if (e->lerpflags & LERP_FINISH)
-			blend = CLAMP (0, (cl.time - e->movelerpstart) / (e->lerpfinish - e->movelerpstart), 1);
+			blend = CLAMP (0.f, (float)(cl.time - e->movelerpstart) / (e->lerpfinish - e->movelerpstart), 1.f);
 		else
-			blend = CLAMP (0, (cl.time - e->movelerpstart) / 0.1, 1);
+			blend = CLAMP (0.f, (float)(cl.time - e->movelerpstart) / 0.1f, 1.f);
 
 		// positional interpolation (translation)
 		VectorSubtract (e->currentorigin, e->previousorigin, d);
